feat: 조합 결과 페이지 UI 구현
