+++ .env
@@ -1,0 +1,1 @@
+VITE_SERVER_API_URL=http://43.203.119.146:8080
