+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@radix-ui/react-switch": "^1.2.5",
     "@tailwindcss/vite": "^4.1.11",
+    "axios": "^1.11.0",
     "clsx": "^2.1.1",
     "lucide-react": "^0.525.0",
     "react": "^19.1.0",
@@ -29,6 +30,7 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.29.0",
+    "@types/axios": "^0.14.4",
     "@types/node": "^24.0.12",
     "@types/react": "^19.1.8",
     "@types/react-dom": "^19.1.6",
