+++ src/App.tsx
@@ -7,9 +7,9 @@ import MainPage from "./pages/MainPage";
 import SignInPage from "./pages/SignInPage";
 import ObjectPage from "./pages/ObjectPage";
 import IngredientPage from "./pages/IngredientPage";
-import CombinationPage from "./pages/CombinationPage";
-import AddCombinationPage from "./pages/AddCombinationPage";
-import CombinationResultPage from "./pages/CombinationResultPage";
+import CombinationPage from "./pages/combination/CombinationPage";
+import AddCombinationPage from "./pages/combination/AddCombinationPage";
+import CombinationResultPage from "./pages/combination/CombinationResultPage";
 import AlarmPage from "./pages/alarm/AlarmPage";
 import MyPage from "./pages/MyPage";
 import AlarmSettingsPage from "./pages/alarm/AlarmSettingsPage";
