+++ src/components/combination/ExpandableProductGroup.tsx
