+++ src/index.css
@@ -3,3 +3,11 @@
   margin: 0;
   padding: 0;
 }
+
+.hide-scrollbar::-webkit-scrollbar {
+  display: none;
+}
+.hide-scrollbar {
+  -ms-overflow-style: none;
+  scrollbar-width: none;
+}
