+++ src/lib/axios.ts
@@ -1,0 +1,16 @@
+import axios from "axios";
+
+const instance = axios.create({
+  baseURL: import.meta.env.VITE_SERVER_API_URL,
+  withCredentials: true,
+});
+
+instance.interceptors.request.use((config) => {
+  const token = localStorage.getItem("accessToken");
+  if (token && config.headers) {
+    config.headers.Authorization = `Bearer ${token}`;
+  }
+  return config;
+});
+
+export default instance;
